requirements.txt
openai-whisper
langchain
langchain-community
pyyaml
json